tests/wpt/mozilla/tests/mozilla/xmlhttprequest_url.html

<!doctype html>
<meta charset="utf-8">
<title>XMLHttpRequest Origin</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
function resolve(path) {
  let a = document.createElement("a");
  a.href = path;
  return a.href;
}

onload = function () {
  let href = window.location.href;
  let request = new XMLHttpRequest();
  let url = resolve("./test.txt");
  request.open('GET', url, true);

  request.onload = test(function () {
    let entries = window.performance.getEntriesByType("resource");
    assert_equals(entries.length, 1);
    assert_equals(entries[0].name, href);
  }, "Performance entries should contain the URL where the XMLHttpRequest originated");

  request.send();
}
</script>
